requirements.txt
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
flask-marshmallow==0.14.0
psycopg2==2.7.5